==> src/3_models.ipynb <==
# Librerías 
import pandas as pd
import polars as pl
import numpy as np
from statistics import mean
from sklearn.model_selection import train_test_split, RandomizedSearchCV, KFold
from sklearn.preprocessing import MinMaxScaler
from sklearn.ensemble import RandomForestClassifier
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report
from pprint import pprint
import pickle

# Lectura 
encig19_inc_raw = pd.read_csv("../data/2_interim/encig19_inc_clean.csv", encoding = "latin1")
variables_contextuales = pd.read_csv("../data/3_final/indicadores_fs.csv", encoding = "utf-8")

# Vista de los DFs 
encig19_inc_raw.head()
encig19_inc_raw.shape

variables_contextuales.head()

# Limpieza

# Se delimitan las variables contextuales al año 2019
variables_contextuales19 = variables_contextuales[variables_contextuales.anio == 2019]

# Se crea el indicador de corrupcion
encig19_inc_clean = encig19_inc_raw
encig19_inc_clean["mean_corrup19"] = encig19_inc_clean.prop_corrup_inc19.mean()
encig19_inc_clean["corrup"] = np.where(encig19_inc_clean.prop_corrup_inc19 > encig19_inc_clean.prop_corrup_inc19.mean(), 1, 0)
encig19_inc_clean.head()
encig19_inc_clean.shape

# Se crea un DF con la proporcion de corrupcion
efipem19_corr_raw = pd.merge(encig19_inc_clean, variables_contextuales19, on = "mun_inegi", how = "left")
efipem19_corr_raw.columns

# Se verifica por NAs
for col in pl.DataFrame(efipem19_corr_raw).get_columns():
    print(f'{col.name} - {col.is_null().sum()}') # Ya no hay NAs

# Se eliminan las columnas extras
efipem19_corr_clean = (
                            pl.DataFrame(efipem19_corr_raw)
                            .lazy()
                            .select(pl.col("mun_inegi", "nom_ent_x", "nom_mun_x", 'mun_tipo', 'graproes', 'pea', 'pe_inac', 'pocupada', 'pdesocup',
       'psinder', 'pder_ss', 'analf', 'sbasc', 'ovsde', 'ovsee', 'ovsae',
       'ovpt', 'vhac', 'pl.5000', 'po2sm', 'im_2020',
       'egresos_total_de_egresos', 'egresos_servicios_personales',
       'egresos_materiales_y_suministros', 'egresos_servicios_generales',
       'egresos_transferencias_asignaciones_subsidios_y_otras_ayudas',
       'egresos_bienes_muebles_inmuebles_e_intangibles',
       'egresos_inversion_publica', 'egresos_deuda_publica',
       'ingresos_total_de_ingresos', 'ingresos_impuestos',
       'ingresos_contribuciones_de_mejoras', 'ingresos_derechos',
       'ingresos_productos', 'ingresos_aprovechamientos',
       'ingresos_participaciones_federales',
       'ingresos_aportaciones_federales_y_estatales',
       'ingresos_financiamiento', 'egresos_disponibilidad_final',
       'egresos_otros_egresos', 'ingresos_otros_ingresos',
       'ingresos_disponibilidad_inicial',
       'egresos_inversiones_financieras_y_otras_provisiones', "corrup").fill_null(0))
                            .collect()
                            .to_dummies(columns = ["mun_tipo"])
                            .drop("mun_tipo_0")
                        )
efipem19_corr_clean

# Se verifica por NAs
for col in efipem19_corr_clean.get_columns():
    print(f'{col.name} - {col.is_null().sum()}') # Ya no hay NAs

indicadores_fs_raw = pd.read_csv("../data/3_final/indicadores_fs.csv")
indicadores_fs_clean = indicadores_fs_raw[indicadores_fs_raw.prop_corrup_inc19.notnull()]

# Remove non-relevant columns
indicadores_fs_clean = indicadores_fs_clean.drop(["nom_ent", "nom_mun", "mun_inegi", "anio", "gm_2020", "prop_corrup_per19", "prop_corrup5_inc19", "ovsae", "ovpt", "po2sm"], axis = 1)

# Remove measure of corruption from other years
indicadores_fs_clean = indicadores_fs_clean.loc[:, ~indicadores_fs_clean.columns.str.endswith("17")]
indicadores_fs_clean = indicadores_fs_clean.loc[:, ~indicadores_fs_clean.columns.str.endswith("21")]

na_cols = []
for col in indicadores_fs_clean.columns:
  na_count = indicadores_fs_clean[col].isna().sum()
  if na_count > 0:
    print(f'{col} - {na_count}')
    na_cols.append(col) 

# Fill NAs
indicadores_fs_clean = indicadores_fs_clean.fillna(value = {"cs_propety_tax_total_revs": indicadores_fs_clean.cs_propety_tax_total_revs.median(),
                                                    "lrs_direct_long_term_debt_pobtot": indicadores_fs_clean.lrs_direct_long_term_debt_pobtot.median(),
                                                    "lrs_debt_service_total_rev": indicadores_fs_clean.lrs_debt_service_total_rev.median()})

indicadores_fs_clean = pd.get_dummies(indicadores_fs_clean, columns = ["mun_tipo"])

# Add feature variable 
indicadores_fs_clean["corrup"] = np.where(indicadores_fs_clean.prop_corrup_inc19 > indicadores_fs_clean.prop_corrup_inc19.mean(), 1, 0)
indicadores_fs_clean = indicadores_fs_clean.drop(["prop_corrup_inc19"], axis = 1)

# Distribution of classes
indicadores_fs_clean.corrup.value_counts()

# Split features from target
X = np.array(indicadores_fs_clean.drop(["corrup"], axis = 1))
Y = np.array(indicadores_fs_clean["corrup"])

# Normalize features
scaler = MinMaxScaler()
X = scaler.fit_transform(X)

# Split Train, test datasets
X_train, X_test, y_train, y_test = train_test_split(X, Y, test_size = 0.2, random_state = 42)

print("X_train:", X_train.shape)
print("X_test:", X_test.shape)

print("y_train:", y_train.shape)
print("y_test:", y_test.shape)

# Compare size of classes in sets
class_train, counts_train = np.unique(y_train, return_counts = True)
class_test, counts_test = np.unique(y_test, return_counts = True)

print("Y train")
print(np.asarray((class_train, counts_train)).T)
print("Y test")
print(np.asarray((class_test, counts_test)).T)


# Se separan las features del target
X = efipem19_corr_clean.drop(["mun_inegi", "nom_ent_x", "nom_mun_x", "corrup"]).to_numpy()
Y = efipem19_corr_clean.select(pl.col("corrup")).to_numpy()

# Se normalizan los datos en X
scaler = MinMaxScaler()
X = scaler.fit_transform(X)

print("Conjunto completo:", efipem19_corr_clean.shape)
print("Features:", X.shape)
print("Target:", Y.shape)

# Se crean los conjuntos de entrenamiento y prueba
X_train, X_test, y_train, y_test = train_test_split(X, Y, test_size = 0.2, random_state = 42)

print("X_train:", X_train.shape)
print("X_test:", X_test.shape)

print("y_train:", y_train.shape)
print("y_test:", y_test.shape)

# Se compara la prevalencia de la clase positiva en ambos conjuntos
class_train, counts_train = np.unique(y_train, return_counts = True)
class_test, counts_test = np.unique(y_test, return_counts = True)

print("Y train")
print(np.asarray((class_train, counts_train)).T)
print("Y test")
print(np.asarray((class_test, counts_test)).T)
# Lucen bien balanceadas

# Cross Validation

## Random Search

# Number of trees in random forest
n_estimators = [int(x) for x in np.linspace(start = 100, stop = 1000, num = 10)]

# Number of features to consider at every split
max_features = ['auto', 'sqrt']
# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(10, 200, num = 20)]
max_depth.append(None)

# Minimum number of samples required to split a node
min_samples_split = [2, 5, 10, 15, 20, 25, 30, 35, 40]
# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20]
# Method of selecting samples for training each tree
bootstrap = [True, False]

# Create the random grid
random_grid = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap}

pprint(random_grid)


# Use the random grid to search for best hyperparameters
# First create the base model to tune
rf = RandomForestClassifier(random_state = 42)

print("Parameters of RF:\n")
pprint(rf.get_params())
# Random search of parameters, using 3 fold cross validation, 
# search across 100 different combinations, and use all available cores
rf_random = RandomizedSearchCV(estimator=rf, param_distributions=random_grid,
                              n_iter = 100, scoring='neg_mean_absolute_error', 
                              cv = 3, verbose=2, random_state=42, n_jobs=2,
                              return_train_score=True)

# Fit the random search model
rf_random.fit(X_train, y_train);

print("BEST PARAMS RF:\n")
print(rf_random.best_params_)

def evaluate(model, test_features, test_labels):
    predictions = model.predict(test_features)
    accuracy = accuracy_score(y_true = test_labels, y_pred = predictions)
    print('Model Performance')
    print('Accuracy = {:0.2f}%.'.format(accuracy*100))
    
    return accuracy

# Evaluate base model
base_model = RandomForestClassifier(n_estimators = 10, random_state = 42)
base_model.fit(X_train, y_train)
base_accuracy = evaluate(base_model, X_test, y_test)

# Evaluate random search model
best_random = rf_random.best_estimator_
random_accuracy = evaluate(best_random, X_test, y_test)

print('Improvement of {:0.2f}%.'.format( 100 * (random_accuracy - base_accuracy) / base_accuracy))

# Validacion cruzada

kf = KFold(n_splits = 10)
CLF_rf = []
f1_rf = []
for train_idx, val_idx in kf.split(X_train):

    # Conjuntos de entrenamiento y validacion
    X_train_kf, y_train_kf = X_train[train_idx], y_train[train_idx]
    X_val, y_val = X_train[val_idx], y_train[val_idx]

    # Random Forest
    clf = RandomForestClassifier(n_estimators = rf_random.best_params_["n_estimators"],
                                 min_samples_split = rf_random.best_params_["min_samples_leaf"],
                                 min_samples_leaf = rf_random.best_params_["min_samples_leaf"],
                                max_features = rf_random.best_params_["max_features"],
                                max_depth = rf_random.best_params_["max_depth"],
                                bootstrap = rf_random.best_params_["bootstrap"]).fit(X_train_kf, y_train_kf)

    # Predicciones
    y_pred = clf.predict(X_val)

    CLF_rf.append(clf)

    print("-"*50)
    print("Resultados del entrenamiento:", len(CLF_rf)-1)
    report = classification_report(y_val, y_pred, digits = 4, output_dict = True)
    f1 = report["macro avg"]["f1-score"]
    f1_rf.append(f1)
    print("Macro avg - F1-score:", round(f1, 4))

# Media total 
print("F1 medio: ", mean(f1_rf))


# Evaluacion final
y_pred_final = CLF_rf[f1_rf.index(max(f1_rf))].predict(X_test)
print(f"RANDOM FOREST EVALUATION\nUsing model with F1 = {max(f1_rf)}")
print(classification_report(y_test, y_pred_final))

# Guardar modelo
pickle.dump(CLF_rf[0], open("../models/rf_90acc.sav", "wb"))
